refactor: Board 와 Board 생성에 관련된 클래스를 포함하는 패키지 생성
