+++ src/test/java/chess/domain/ChessGameTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.domain.piece.Color;
+import chess.domain.board.TestOtherTeamBoard;
 import chess.domain.position.Square;
 
 public class ChessGameTest {
@@ -13,8 +13,7 @@ public class ChessGameTest {
     @Test
     @DisplayName("흰 말로 시작하지 않으면 에러를 반환한다")
     void errorTurn_Start() {
-        Board chessBoard = new Board(new TestOtherTeamBoard());
-        ChessGame chessGame = new ChessGame(chessBoard, Color.WHITE);
+        ChessGame chessGame = new ChessGame(new TestOtherTeamBoard());
         assertThatThrownBy(() -> chessGame.move(new Square("d4"), new Square("e5")))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 순서 지키시지?!");
@@ -23,8 +22,7 @@ public class ChessGameTest {
     @Test
     @DisplayName("흰 말 다음 검은 말 순서로 진행하지 않으면 에러를 반환한다")
     void errorTurn() {
-        Board chessBoard = new Board(new TestOtherTeamBoard());
-        ChessGame chessGame = new ChessGame(chessBoard, Color.WHITE);
+        ChessGame chessGame = new ChessGame(new TestOtherTeamBoard());
         chessGame.move(new Square("c3"), new Square("d4"));
         assertThatThrownBy(() -> chessGame.move(new Square("d4"), new Square("e5")))
                 .isInstanceOf(IllegalArgumentException.class)
